test: move async dispatch test
